@@ -29,39 +42,53 @@ generate-setup-file = false
 
 # Requirements
 [tool.poetry.dependencies]
-python = "^3.6"
+python = "^3.7"
 
-poetry-core = "^1.1.0a6"
-cleo = "^1.0.0a4"
-crashtest = "^0.3.0"
-requests = "^2.18"
+poetry-core = "^1.1.0b1"
+poetry-plugin-export = "^1.0.4"
+cachecontrol = { version = "^0.12.9", extras = ["filecache"] }
 cachy = "^0.3.0"
-requests-toolbelt = "^0.9.1"
-cachecontrol = { version = "^0.12.4", extras = ["filecache"] }
-pkginfo = "^1.5"
+cleo = "^1.0.0a5"
+crashtest = "^0.3.0"
+entrypoints = "^0.4"
 html5lib = "^1.0"
+importlib-metadata = { version = "^4.4", python = "<3.10" }
+# keyring uses calver, so version is unclamped
+keyring = ">=21.2.0"
+# packaging uses calver, so version is unclamped
+packaging = ">=20.4"
+pexpect = "^4.7.0"
+pkginfo = "^1.5"
+platformdirs = "^2.5.2"
+requests = "^2.18"
+requests-toolbelt = "^0.9.1"
 shellingham = "^1.1"
 tomlkit = ">=0.7.0,<1.0.0"
-pexpect = "^4.7.0"
-packaging = "^20.4"
-# temporarily clamped due to https://github.com/pypa/pip/issues/9953
-virtualenv = ">=20.4.3,<20.4.5"
-keyring = ">=21.2.0"
-entrypoints = "^0.3"
-importlib-metadata = {version = "^1.6.0", python = "<3.8"}
-dataclasses = {version = "^0.8", python = "~3.6"}
+# exclude 20.4.5 - 20.4.6 due to https://github.com/pypa/pip/issues/9953
+virtualenv = "(>=20.4.3,<20.4.5 || >=20.4.7)"
 xattr = {version = "^0.9.7", markers = "sys_platform == 'darwin'"}
+urllib3 = "^1.26.0"
+dulwich = "^0.20.35"
 
 [tool.poetry.dev-dependencies]
-pytest = "^6.2"
-pytest-cov = "^2.8"
+tox = "^3.18"
+pytest = "^7.1"
+pytest-cov = "^3.0"
 pytest-mock = "^3.5"
-pre-commit = { version = "^2.6", python = "^3.6.1" }
-tox = "^3.0"
+pytest-randomly = "^3.10"
 pytest-sugar = "^0.9"
-httpretty = "^1.0"
-zipp = { version = "^3.4", python = "<3.8"}
+pytest-xdist = { version = "^2.5", extras = ["psutil"] }
+pre-commit = "^2.6"
 deepdiff = "^5.0"
+httpretty = "^1.0"
+typing-extensions = { version = "^4.0.0", python = "<3.8" }
+zipp = { version = "^3.4", python = "<3.8" }
+flatdict = "^4.0.1"
+mypy = ">=0.960"
+types-entrypoints = ">=0.3.7"
+types-html5lib = ">=1.1.7"
+types-jsonschema = ">=4.4.4"
+types-requests = ">=2.27.11"
 
 [tool.poetry.scripts]
 poetry = "poetry.console.application:main"
