@@ -1,26 +1,41 @@
+from __future__ import annotations
+
 import os
+
+from typing import TYPE_CHECKING
 
 import pytest
 
-from poetry.utils.password_manager import KeyRing
-from poetry.utils.password_manager import KeyRingError
 from poetry.utils.password_manager import PasswordManager
+from poetry.utils.password_manager import PoetryKeyring
+from poetry.utils.password_manager import PoetryKeyringError
 
 
-def test_set_http_password(config, with_simple_keyring, dummy_keyring):
+if TYPE_CHECKING:
+    from pytest_mock import MockerFixture
+
+    from tests.conftest import Config
+    from tests.conftest import DummyBackend
+
+
+def test_set_http_password(
+    config: Config, with_simple_keyring: None, dummy_keyring: DummyBackend
+):
     manager = PasswordManager(config)
 
     assert manager.keyring.is_available()
     manager.set_http_password("foo", "bar", "baz")
 
-    assert "baz" == dummy_keyring.get_password("poetry-repository-foo", "bar")
+    assert dummy_keyring.get_password("poetry-repository-foo", "bar") == "baz"
 
     auth = config.get("http-basic.foo")
-    assert "bar" == auth["username"]
+    assert auth["username"] == "bar"
     assert "password" not in auth
 
 
-def test_get_http_auth(config, with_simple_keyring, dummy_keyring):
+def test_get_http_auth(
+    config: Config, with_simple_keyring: None, dummy_keyring: DummyBackend
+):
     dummy_keyring.set_password("poetry-repository-foo", "bar", "baz")
     config.auth_config_source.add_property("http-basic.foo", {"username": "bar"})
     manager = PasswordManager(config)
@@ -28,11 +43,13 @@ def test_get_http_auth(config, with_simple_keyring, dummy_keyring):
     assert manager.keyring.is_available()
     auth = manager.get_http_auth("foo")
 
-    assert "bar" == auth["username"]
-    assert "baz" == auth["password"]
+    assert auth["username"] == "bar"
+    assert auth["password"] == "baz"
 
 
-def test_delete_http_password(config, with_simple_keyring, dummy_keyring):
+def test_delete_http_password(
+    config: Config, with_simple_keyring: None, dummy_keyring: DummyBackend
+):
     dummy_keyring.set_password("poetry-repository-foo", "bar", "baz")
     config.auth_config_source.add_property("http-basic.foo", {"username": "bar"})
     manager = PasswordManager(config)
@@ -44,7 +61,9 @@ def test_delete_http_password(config, with_simple_keyring, dummy_keyring):
     assert config.get("http-basic.foo") is None
 
 
-def test_set_pypi_token(config, with_simple_keyring, dummy_keyring):
+def test_set_pypi_token(
+    config: Config, with_simple_keyring: None, dummy_keyring: DummyBackend
+):
     manager = PasswordManager(config)
 
     assert manager.keyring.is_available()
@@ -52,18 +71,22 @@ def test_set_pypi_token(config, with_simple_keyring, dummy_keyring):
 
     assert config.get("pypi-token.foo") is None
 
-    assert "baz" == dummy_keyring.get_password("poetry-repository-foo", "__token__")
+    assert dummy_keyring.get_password("poetry-repository-foo", "__token__") == "baz"
 
 
-def test_get_pypi_token(config, with_simple_keyring, dummy_keyring):
+def test_get_pypi_token(
+    config: Config, with_simple_keyring: None, dummy_keyring: DummyBackend
+):
     dummy_keyring.set_password("poetry-repository-foo", "__token__", "baz")
     manager = PasswordManager(config)
 
     assert manager.keyring.is_available()
-    assert "baz" == manager.get_pypi_token("foo")
+    assert manager.get_pypi_token("foo") == "baz"
 
 
-def test_delete_pypi_token(config, with_simple_keyring, dummy_keyring):
+def test_delete_pypi_token(
+    config: Config, with_simple_keyring: None, dummy_keyring: DummyBackend
+):
     dummy_keyring.set_password("poetry-repository-foo", "__token__", "baz")
     manager = PasswordManager(config)
 
@@ -73,18 +96,22 @@ def test_delete_pypi_token(config, with_simple_keyring, dummy_keyring):
     assert dummy_keyring.get_password("poetry-repository-foo", "__token__") is None
 
 
-def test_set_http_password_with_unavailable_backend(config, with_fail_keyring):
+def test_set_http_password_with_unavailable_backend(
+    config: Config, with_fail_keyring: None
+):
     manager = PasswordManager(config)
 
     assert not manager.keyring.is_available()
     manager.set_http_password("foo", "bar", "baz")
 
     auth = config.get("http-basic.foo")
-    assert "bar" == auth["username"]
-    assert "baz" == auth["password"]
+    assert auth["username"] == "bar"
+    assert auth["password"] == "baz"
 
 
-def test_get_http_auth_with_unavailable_backend(config, with_fail_keyring):
+def test_get_http_auth_with_unavailable_backend(
+    config: Config, with_fail_keyring: None
+):
     config.auth_config_source.add_property(
         "http-basic.foo", {"username": "bar", "password": "baz"}
     )
@@ -93,11 +120,13 @@ def test_get_http_auth_with_unavailable_backend(config, with_fail_keyring):
     assert not manager.keyring.is_available()
     auth = manager.get_http_auth("foo")
 
-    assert "bar" == auth["username"]
-    assert "baz" == auth["password"]
+    assert auth["username"] == "bar"
+    assert auth["password"] == "baz"
 
 
-def test_delete_http_password_with_unavailable_backend(config, with_fail_keyring):
+def test_delete_http_password_with_unavailable_backend(
+    config: Config, with_fail_keyring: None
+):
     config.auth_config_source.add_property(
         "http-basic.foo", {"username": "bar", "password": "baz"}
     )
@@ -109,24 +138,30 @@ def test_delete_http_password_with_unavailable_backend(config, with_fail_keyring
     assert config.get("http-basic.foo") is None
 
 
-def test_set_pypi_token_with_unavailable_backend(config, with_fail_keyring):
+def test_set_pypi_token_with_unavailable_backend(
+    config: Config, with_fail_keyring: None
+):
     manager = PasswordManager(config)
 
     assert not manager.keyring.is_available()
     manager.set_pypi_token("foo", "baz")
 
-    assert "baz" == config.get("pypi-token.foo")
+    assert config.get("pypi-token.foo") == "baz"
 
 
-def test_get_pypi_token_with_unavailable_backend(config, with_fail_keyring):
+def test_get_pypi_token_with_unavailable_backend(
+    config: Config, with_fail_keyring: None
+):
     config.auth_config_source.add_property("pypi-token.foo", "baz")
     manager = PasswordManager(config)
 
     assert not manager.keyring.is_available()
-    assert "baz" == manager.get_pypi_token("foo")
+    assert manager.get_pypi_token("foo") == "baz"
 
 
-def test_delete_pypi_token_with_unavailable_backend(config, with_fail_keyring):
+def test_delete_pypi_token_with_unavailable_backend(
+    config: Config, with_fail_keyring: None
+):
     config.auth_config_source.add_property("pypi-token.foo", "baz")
     manager = PasswordManager(config)
 
@@ -136,29 +171,57 @@ def test_delete_pypi_token_with_unavailable_backend(config, with_fail_keyring):
     assert config.get("pypi-token.foo") is None
 
 
-def test_keyring_raises_errors_on_keyring_errors(mocker, with_fail_keyring):
-    mocker.patch("poetry.utils.password_manager.KeyRing._check")
+def test_keyring_raises_errors_on_keyring_errors(
+    mocker: MockerFixture, with_fail_keyring: None
+):
+    mocker.patch("poetry.utils.password_manager.PoetryKeyring._check")
 
-    key_ring = KeyRing("poetry")
-    with pytest.raises(KeyRingError):
+    key_ring = PoetryKeyring("poetry")
+    with pytest.raises(PoetryKeyringError):
         key_ring.set_password("foo", "bar", "baz")
 
-    with pytest.raises(KeyRingError):
+    with pytest.raises(PoetryKeyringError):
         key_ring.get_password("foo", "bar")
 
-    with pytest.raises(KeyRingError):
+    with pytest.raises(PoetryKeyringError):
         key_ring.delete_password("foo", "bar")
 
 
-def test_keyring_with_chainer_backend_and_not_compatible_only_should_be_unavailable(
-    with_chained_keyring,
+def test_keyring_with_chainer_backend_and_fail_keyring_should_be_unavailable(
+    with_chained_fail_keyring: None,
 ):
-    key_ring = KeyRing("poetry")
+    key_ring = PoetryKeyring("poetry")
 
     assert not key_ring.is_available()
 
 
-def test_get_http_auth_from_environment_variables(environ, config, with_simple_keyring):
+def test_keyring_with_chainer_backend_and_null_keyring_should_be_unavailable(
+    with_chained_null_keyring: None,
+):
+    key_ring = PoetryKeyring("poetry")
+
+    assert not key_ring.is_available()
+
+
+def test_null_keyring_should_be_unavailable(
+    with_null_keyring: None,
+):
+    key_ring = PoetryKeyring("poetry")
+
+    assert not key_ring.is_available()
+
+
+def test_fail_keyring_should_be_unavailable(
+    with_fail_keyring: None,
+):
+    key_ring = PoetryKeyring("poetry")
+
+    assert not key_ring.is_available()
+
+
+def test_get_http_auth_from_environment_variables(
+    environ: None, config: Config, with_simple_keyring: None
+):
     os.environ["POETRY_HTTP_BASIC_FOO_USERNAME"] = "bar"
     os.environ["POETRY_HTTP_BASIC_FOO_PASSWORD"] = "baz"
 
@@ -166,5 +229,5 @@ def test_get_http_auth_from_environment_variables(environ, config, with_simple_k
 
     auth = manager.get_http_auth("foo")
 
-    assert "bar" == auth["username"]
-    assert "baz" == auth["password"]
+    assert auth["username"] == "bar"
+    assert auth["password"] == "baz"
